Begin to put in scaffolding to handle ascii better.
AsciiScreen should handle the space it's given to render better based on
the size of a field, size of window, number of characters to draw at a a
time.

TODO comments everywhere.

Eventually we'll want to do more complex things with characters, like
make them flash or give them borders or whatever. Some scaffolding has
started to be put into place to handle this.

AsciiScreen will remain completely agnostic to whatever framework is
using it. Instead of rendering anything itself, it will pass off all the
necessary bits of information to a function it is given and let whoever is
telling it to render handle actually rendering strings.