Remove font handling; Todos
Added a bunch of TODO comments for future improvements.
Please ignore them in the main comment block for the
class, just trying to make sure I'm not forgetting any
ideas.

Whether or not an AsciiScreen is in UI only mode is now
public information.

Remove any and all font handling code. Due to the
vastly different way that fonts are handled between
frameworks, it doesn't make sense to attempt to handle
anything about fonts in AsciiScreen. External code will
need to load its own fonts and then render any
characters with those loaded fonts.

Removed the pooled proxy items. There is no
guarantee that external code will release a hold on
them when an item is removed, so it's safer to just
return a new one whenever a new item is placed.

A placed item can now have any object associated with
it as user data.

Function passed to render now gets passed a
CharDrawData object instead of a bunch of other bits of
data. This object should hopefully make it easy to
draw the necessary text, as well as make API changes
easy as they can just be added to the object.