Implement movement action
Currently purely reuses remove/placement actions which necessitates allocating
two extra objects per movement. For an initial version this is fine, but it should
definitely be changed eventually to have some static methods that all three
actions use instead of needing to allocated objects to access the common
functionality of removing/adding items to a layer.